[systemd] Made .deb package systemd compliant
Without breaking compatibility with SysVInit, all of it in a single
package (maybe in a near future making separate package using runit in
Omnibus would be a better option).

Some cleanup has also been made (replacing which by command -v to be
POSIX compliant, moving a chmod call from the postinst script to the
package extraction phase, adding one or two echoes...)

Note about systemd call to a start_agent.sh script
--------------------------------------------------

The custom PATH used by the agent is now setup in a separate script when
using systemd. The following exec call ensure that we keep the same
forking logic as before (and is also cleaner than calling /bin/bash
directly in the ExecStart directive of the unit file).

Moved service disabling to prerm
--------------------------------

Disabling the datadog-agent service has been moved to the prerm script.
Indeed, if we do it after removal, update-rc.d doesn't realize it has to
delete a systemd service, tries to delete a SysVInit one and fails with
exit code 1.