Add RPM signing features
The key has to be shared with the docker image (the whole /root/.gnupg
folder could be synced with the ~/.gnupg of the build machine for
instance). The docker container must also be passed an
RPM_SIGNING_PASSPHRASE containing the signing passphrase corresponding
to the desired key. The key owner name and email string must match
EXACTLY the name and e-mail string of the key owner (displayed when
doing a gpg --list-secret-keys.