[osx] Migration to omnibus 4
The project has been made compliant with the OSX Omnibus build. Builds
are now on the same branch for OSX and Linux and can be triggered using
the same set of commands. Please have a look at the README.md file in
order to make sure you have all the requirements installed on your Mac.