[osx] start agent at login automatically
And link all conf files to ~/.datadog-agent
Also (for simplification reasons):
the app is totally destroyed (except configuration, ofc) during the preinstall
script, to avoid any error and be in the same state for upgrade install for
postinstall.
Why do this ? Avoid Mac-behavior dependencies for the upgrade, it doesn't
always behave as it should.