[windows] Relocate the build to have pip.exe work
After modyfing our fork of Omnibus4 a little (just adding the right
quotes in the right places, that wasn't too hard...) we can build in the
Program Files folder, which we do. This allows all the Python scripts
(pip mainly) to work seamlessly, out of the box if the agent is
installed to the defaut directory. If not, the `pip-script.py` (and
equivalent `.py` files for other scripts) will still keep on working as
expected when launched with the embedded `python.exe`.

This also contains some cleanups.