Fix dd-agent CLI on Linux
The `/usr/bin/dd-agent command haven't worked
anymore since... a big while. I suspect that this is due to a slight
variation in the behaviour of the `extra_package_file` between Omnibus 3
and Omnibus 4 but didn't really have time to dig more into it.

This fix just moves the creation of this file to the `postinst` script
which fixes the issue in a rather simple way. This same file also get
deleted in the `postrm` script so that the `dd-agent` commands simply
don't exist anymore after removing the agent.