[linux] Fixed dd-agent CLI: Part II
The order in with package scripts get executed during updates with YUM
led to an unexpected behaviour. After a couple of minutes of
facepalming, a decent fix has been implemented. There's also a README in
the `package-scripts/datadog-agent` folder that details a bit more how
updates are handled by APT and YUM, and there's a couple of rather
important differences to pay attention to when updating these scripts :)