[prerm][python] Better handling of the pyc/pyo files removal on RPM
With RPM, if we remove the pyc/pyo files in the `prerm` script, when
the package is upgraded, we end up removing the files that are listed
in the `.py_compiled_files.txt` file of the _new_ version of the
package because the files of the new package get copied _before_ the
`prerm` script of the old package is run.

To avoid this, on RPM, we clean the pyc/pyo files in the prerm script
only when the package is removed. When the package is upgraded we do
the cleaning in the preinst script.

Behaviour unchanged for DEBs.

Also, updated docs.