[postinst] Make `dd-agent` user and group own `/opt/datadog-agent`
Users of the Agent can then give no permissions at all to other users
on the `/opt/datadog-agent` directory (which is the home directory of
`dd-agent`), and the Agent can still work fine and not break the users'
changes on upgrade.

Also, on debian, set the home directory of `dd-agent` to
`/opt/datadog-agent` directly on the `adduser` call, for the sake of
brevity.