fixes integrations
This started out as a small bug fix and balooned as a found other bugs.

This fixes a small syntax error, first. It removes a few debugging lines, such as when I was ls -l the integrations directory.

It fixes adding the config files, it changes the sdk directory to integrations, rather than 3rd-party, and it makes the pip install more fault tolerant.

Finally, it (hopefully) fixes the license file conflict.