[deb] Stop adding new key (#237)
The new key is now used to sign the repo. No point in adding it in the
preinst script since a user won't be able to download the package
if the key's not trusted on the system.

Also, remove dep on `gnupg` that came from the `apt-key` usage in
preinst script.